Turn addon-dir into an absolute path (fixes #483)
The addonDir option seems to only work fine with absolute paths. This is
fine as long as jpm only uses those internally and for tests, but it
breaks the `jpm xpi --addon-dir ./my/addon` command. The easiest fix is
to just run a transform function on that CLI option that makes it absolute.